refactor #34 : 컨벤션에 따른 파일명 변경(index제거)
